requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
boto3==1.42.38
botocore==1.42.38
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.4
dnspython==2.8.0
easydict==1.13
fastapi==0.128.0
flatbuffers==25.12.19
fonttools==4.61.1
h11==0.16.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
insightface==0.7.3
jmespath==1.1.0
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.8
ml_dtypes==0.5.4
mongoengine==0.29.1
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
onnx==1.20.1
onnxruntime==1.23.2
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.90
packaging==26.0
pillow==12.1.0
prettytable==3.17.0
protobuf==6.33.5
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
s3transfer==0.16.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
simsimd==6.5.12
six==1.17.0
starlette==0.50.0
stringzilla==4.6.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2026.1.28
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
wcwidth==0.5.2
